+++ src/migrations/m220222_122159_full_names.php
@@ -21,7 +21,7 @@ class m220222_122159_full_names extends Migration
 
         $this->update(Table::USERS, [
             'fullName' => new Expression("TRIM(CONCAT([[firstName]], ' ', [[lastName]]))"),
-        ], new Expression("TRIM(CONCAT([[firstName]], [[lastName]])) <> ''"), [], false);
+        ], new Expression("TRIM(CONCAT([[firstName]], [[lastName]])) <> ''"), updateTimestamp: false);
 
         return true;
     }