+++ src/web/AssetManager.php
@@ -83,7 +83,7 @@ class AssetManager extends \yii\web\AssetManager
             Db::upsert(Table::RESOURCEPATHS, [
                 'hash' => $hash,
                 'path' => $alias,
-            ], true, [], false);
+            ]);
         } catch (DbException | DbConnectException $e) {
             // Craft is either not installed or not updated to 3.0.3+ yet
         }